+++ samples/Conventions/Program.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (c) Nate McMaster.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+public class Program
+{
+    static void Main(string[] args)
+    {
+        DependencyInjectionProgram.Main(args);
+    }
+}