+++ samples/Prompt/Program.cs
@@ -1,4 +1,7 @@
-﻿using System;
+﻿// Copyright (c) Nate McMaster.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
 using McMaster.Extensions.CommandLineUtils;
 
 class Program