+++ src/CommandLineUtils/Abstractions/IModelAccessor.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) Nate McMaster.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using System;
+
 namespace McMaster.Extensions.CommandLineUtils.Abstractions
 {
     /// <summary>
@@ -8,6 +10,12 @@ namespace McMaster.Extensions.CommandLineUtils.Abstractions
     /// </summary>
     public interface IModelAccessor
     {
+        /// <summary>
+        /// Gets the type of the model.
+        /// </summary>
+        /// <returns>The type.</returns>
+        Type GetModelType();
+
         /// <summary>
         /// Gets the model.
         /// </summary>