+++ src/CommandLineUtils/Internal/CommandLineValidationContext.cs
@@ -18,49 +18,9 @@ namespace McMaster.Extensions.CommandLineUtils
         }
 
         public ValidationContext Create(CommandArgument argument)
-            => new ValidationContext(argument, new ServiceProvider(_app), null);
+            => new ValidationContext(argument, _app, null);
 
         public ValidationContext Create(CommandOption option)
-            => new ValidationContext(option, new ServiceProvider(_app), null);
-
-        private sealed class ServiceProvider : IServiceProvider
-        {
-            private readonly CommandLineApplication _parent;
-
-            public ServiceProvider(CommandLineApplication parent)
-            {
-                _parent = parent;
-            }
-
-            public object GetService(Type serviceType)
-            {
-                if (serviceType == typeof(CommandLineApplication))
-                {
-                    return _parent;
-                }
-
-                if (serviceType == typeof(IConsole))
-                {
-                    return _parent._context.Console;
-                }
-
-                if (serviceType == typeof(IEnumerable<CommandOption>))
-                {
-                    return _parent.GetOptions();
-                }
-
-                if (serviceType == typeof(IEnumerable<CommandArgument>))
-                {
-                    return _parent.Arguments;
-                }
-
-                if (serviceType == typeof(CommandLineContext))
-                {
-                    return _parent._context;
-                }
-
-                return null;
-            }
-        }
+            => new ValidationContext(option, _app, null);
     }
 }